halfway through refactoring function, module and file locations...but so far it's working
